+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog] and this project adheres to
 ## [Unreleased]
 ### Fixed
 - Undo/Redo for Pick Residential Area action.
+- Undo/Redo for Batch Orthogonalize Building action.
 
 ## [0.2.3] - 2018-05-08
 ### Fixed